+++ network/network/src/integrationTest/java/bisq/network/p2p/node/BaseNodesByIdTest.java
@@ -25,8 +25,8 @@ import bisq.network.p2p.node.authorization.AuthorizationTokenType;
 import bisq.network.p2p.node.network_load.NetworkLoadSnapshot;
 import bisq.network.p2p.node.transport.TransportService;
 import bisq.network.p2p.services.peergroup.BanList;
-import bisq.network.p2p.services.peergroup.keepalive.Ping;
-import bisq.network.p2p.services.peergroup.keepalive.Pong;
+import bisq.network.p2p.services.peergroup.keep_alive.Ping;
+import bisq.network.p2p.services.peergroup.keep_alive.Pong;
 import bisq.persistence.PersistenceService;
 import bisq.security.keys.KeyBundleService;
 import bisq.security.pow.equihash.EquihashProofOfWorkService;
