+++ network/network/src/main/java/bisq/network/p2p/services/peergroup/keep_alive/Ping.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.network.p2p.services.peergroup.keepalive;
+package bisq.network.p2p.services.peergroup.keep_alive;
 
 import bisq.network.p2p.message.EnvelopePayloadMessage;
 import bisq.network.p2p.message.Request;
